requirements.txt
astroid==2.15.8
certifi==2022.12.7
charset-normalizer==2.1.1
colorama==0.4.6
contourpy==1.1.1
cycler==0.11.0
dill==0.3.7
filelock==3.9.0
fonttools==4.42.1
fsspec==2023.4.0
idna==3.4
isort==5.12.0
Jinja2==3.1.2
kiwisolver==1.4.5
lazy-object-proxy==1.9.0
MarkupSafe==2.1.2
matplotlib==3.8.0
mccabe==0.7.0
mpmath==1.2.1
networkx==3.0rc1
numpy==1.24.1
opencv-python-headless==4.8.0.76
packaging==23.1
Pillow==9.3.0
platformdirs==3.10.0
pylint==2.17.6
pyparsing==3.1.1
python-dateutil==2.8.2
requests==2.28.1
six==1.16.0
sympy==1.11.1
tomlkit==0.12.1
torch==2.2.0.dev20230922+cu121
torchaudio==2.2.0.dev20230925+cu121
torchvision==0.17.0.dev20230925+cu121
tqdm==4.66.1
typing_extensions==4.4.0
urllib3==1.26.13
wrapt==1.15.0
